backend/requirements.txt
fastapi
uvicorn[standard]
pydub
numpy
python-socketio[asgi]
audioop-lts